pyproject.toml
[tool.poetry]
name = "corpus-unpdf"
version = "0.0.1"
description = "Methods to extract pdf to text"
authors = ["Marcelino G. Veloso III <mars@veloso.one>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^0.21"
pdfplumber = "^0.7.6"
pillow = "^9.4.0"
opencv-python = "^4.7.0.68"
pytesseract = "^0.3.10"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
pre-commit = "^2.21"
jupyter = "^1.0.0"
mkdocs = "^1.4.2"
mkdocstrings = { extras = ["python"], version = "^0.20.0" }
mkdocs-material = "^9.0.6"

[tool.pytest]
minversion = "7.2"
addopts = "-ra -q --cov=tests/"
testpaths = ["tests"]

[tool.ruff]
ignore = ["F401"]

[tool.black]
target-version = ['py311']
line-length = 79
include = '.pyi?$'
exclude = '''
/(
    .git
    | .hg
    | .mypy_cache
    | .tox
    | .venv
    | _build
    | buck-out
    | build
    | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 79
include_trailing_comma = 'True'

[build-system]
requires = ["poetry-core>=1.3.2"]
build-backend = "poetry.core.masonry.api"
